keep model in types
