+++ packages/providers/google/src/studio/index.ts
@@ -26,11 +26,8 @@ import { streamConverter } from "@llamaindex/core/utils";
 import { wrapLLMEvent } from "@llamaindex/core/decorator";
 import type { JSONObject } from "@llamaindex/core/global";
 import { randomUUID } from "@llamaindex/env";
-import {
-  DEFAULT_GEMINI_PARAMS,
-  GEMINI_MODEL,
-  SUPPORT_TOOL_CALL_MODELS,
-} from "../constants";
+import { DEFAULT_GEMINI_PARAMS, SUPPORT_TOOL_CALL_MODELS } from "../constants";
+import { GEMINI_MODEL } from "../types";
 import {
   mapChatMessagesToGoogleFunctions,
   mapChatMessagesToGoogleMessages,
